pyproject.toml
[build-system]
requires = ["setuptools>=61,<69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sam-faces"
version = "1.0.0"
description = "Face recognition and identity memory for AI assistants"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Sam Cox", email = "sam@jasonacox.com"}
]
keywords = ["face-recognition", "ai", "agent", "identity", "memory", "openclaw"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "face-recognition>=1.3.0",
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black",
    "build",
    "twine",
]

[project.scripts]
sam-faces = "sam_faces.cli:main"

[project.urls]
Homepage = "https://github.com/jasonacox-sam/sam-faces"
Documentation = "https://github.com/jasonacox-sam/sam-faces"
Repository = "https://github.com/jasonacox-sam/sam-faces"

[tool.setuptools.packages.find]
where = ["."]
include = ["sam_faces*"]
